+++ components/api_server/tests/initialization/test_database.py
@@ -3,7 +3,7 @@
 import pathlib
 from unittest.mock import Mock, mock_open, patch
 
-from initialization.database import init_database
+from initialization.database import init_database, perform_migrations
 
 from tests.base import DataModelTestCase
 
@@ -61,3 +61,95 @@ class DatabaseInitTest(DataModelTestCase):
             self.init_database('{"change": "yes"}', False)
         self.database.datamodels.insert_one.assert_called_once()
         self.database.reports_overviews.insert_one.assert_called_once()
+
+
+class DatabaseMigrationsTest(DataModelTestCase):
+    """Unit tests for the database migration."""
+
+    def setUp(self):
+        """Override to set up the database fixture."""
+        self.database = Mock()
+
+    def test_change_accessibility_violations_to_violations_without_reports(self):
+        """Test that the migration succeeds without reports."""
+        self.database.reports.find.return_value = []
+        perform_migrations(self.database)
+        self.database.reports.replace_one.assert_not_called()
+
+    def test_change_accessibility_violations_to_violations_when_report_has_no_accessibility_metrics(self):
+        """Test that the migration succeeds wtih reports, but without accessibility metrics."""
+        self.database.reports.find.return_value = [
+            {"report_uuid": "report_uuid", "subjects": {"subject_uuid": {"metrics": {"metric_uuid": {"type": "loc"}}}}}
+        ]
+        perform_migrations(self.database)
+        self.database.reports.replace_one.assert_not_called()
+
+    def test_change_accessibility_violations_to_violations_when_report_has_an_accessibility_metric(self):
+        """Test that the migration succeeds with an accessibility metric."""
+        self.database.reports.find.return_value = [
+            {
+                "_id": "id",
+                "report_uuid": "report_uuid",
+                "subjects": {"subject_uuid": {"metrics": {"metric_uuid": {"type": "accessibility"}}}},
+            },
+        ]
+        perform_migrations(self.database)
+        self.database.reports.replace_one.assert_called_once_with(
+            {"_id": "id"},
+            {
+                "report_uuid": "report_uuid",
+                "subjects": {
+                    "subject_uuid": {
+                        "metrics": {
+                            "metric_uuid": {
+                                "type": "violations",
+                                "name": "Accessibility violations",
+                                "unit": "accessibility violations",
+                            }
+                        }
+                    }
+                },
+            },
+        )
+
+    def test_change_accessibility_violations_to_violations_when_report_has_accessibility_metric_and_other_types(self):
+        """Test that the migration succeeds with an accessibility metric and other metric types."""
+        self.database.reports.find.return_value = [
+            {
+                "_id": "id",
+                "report_uuid": "report_uuid",
+                "subjects": {
+                    "subject_uuid": {
+                        "metrics": {
+                            "metric_uuid": {"type": "accessibility"},
+                            "metric_uuid2": {"type": "violations"},
+                            "metric_uuid3": {"type": "security_warnings"},
+                        }
+                    }
+                },
+            }
+        ]
+        perform_migrations(self.database)
+        self.database.reports.replace_one.assert_called_once_with(
+            {"_id": "id"},
+            {
+                "report_uuid": "report_uuid",
+                "subjects": {
+                    "subject_uuid": {
+                        "metrics": {
+                            "metric_uuid": {
+                                "type": "violations",
+                                "name": "Accessibility violations",
+                                "unit": "accessibility violations",
+                            },
+                            "metric_uuid2": {
+                                "type": "violations",
+                            },
+                            "metric_uuid3": {
+                                "type": "security_warnings",
+                            },
+                        }
+                    }
+                },
+            },
+        )